Accessor class and controller class implemented
Routers now use two different classes to remove code repetition from files.
