+++ routers/drugs-Router.js
@@ -1,181 +1,30 @@
 import { Router } from "express";
 import Model from "../models/Model.js";
 import modelConfig from "../models/drugs-model.js";
+import Accessor from "../accessor/Accessor.js";
 import database from "../database.js";
+import Controller from "../controller/Controller.js";
 
-const router = Router();
 // Model -------------------------------------
 
 const model = new Model(modelConfig);
 
 // Data accessors -------------------------------------
 
-const create = async (record) => {
-  try {
-    const { sql, data } = model.buildCreateQuery(record);
-    const status = await database.query(sql, data);
-    const { isSuccess, result, message } = await read(status[0].insertId, null);
-
-    return isSuccess
-      ? {
-          isSuccess: true,
-          result: result,
-          message: "Record successfully recovered",
-        }
-      : {
-          isSuccess: false,
-          result: null,
-          message: `Failed to recover the inserted record ${message}`,
-        };
-  } catch (error) {
-    return {
-      isSuccess: false,
-      result: null,
-      message: `Failed to execute query: ${error.message}`,
-    };
-  }
-};
-
-const read = async (id, variant) => {
-  try {
-    const { sql, data } = model.buildReadQuery(id, variant);
-    const [result] = await database.query(sql, data);
-    return result.length === 0
-      ? { isSuccess: false, result: null, message: "No record(s) found" }
-      : {
-          isSuccess: true,
-          result: result,
-          message: "Record(s) successfully recovered",
-        };
-  } catch (error) {
-    return {
-      isSuccess: false,
-      result: null,
-      message: `Failed to execute query: ${error.message}`,
-    };
-  }
-};
-
-const update = async (record, id) => {
-  try {
-    const { sql, data } = model.buildUpdateQuery(record, id);
-    const status = await database.query(sql, data);
-
-    if (status[0].affectedRows === 0)
-      return {
-        isSuccess: false,
-        result: null,
-        message: `Failed to update record: no rows affected`,
-      };
-
-    const { isSuccess, result, message } = await read(id, null);
-
-    return isSuccess
-      ? {
-          isSuccess: true,
-          result: result,
-          message: "Record successfully recovered",
-        }
-      : {
-          isSuccess: false,
-          result: null,
-          message: `Failed to recover the updated record ${message}`,
-        };
-  } catch (error) {
-    return {
-      isSuccess: false,
-      result: null,
-      message: `Failed to execute query: ${error.message}`,
-    };
-  }
-};
-
-const _delete = async (id) => {
-  try {
-    const { sql, data } = model.buildDeleteQuery(id);
-    const status = await database.query(sql, data);
-    return status[0].affectedRows === 0
-      ? {
-          isSuccess: false,
-          result: null,
-          message: `Failed to delete record ${id}`,
-        }
-      : {
-          isSuccess: true,
-          result: null,
-          message: "Record successfully deleted",
-        };
-  } catch (error) {
-    return {
-      isSuccess: false,
-      result: null,
-      message: `Failed to execute query: ${error.message}`,
-    };
-  }
-};
+const accessor = new Accessor(model, database);
 
 // Controllers -------------------------------------
 
-const getDrugsController = async (req, res, variant) => {
-  const id = req.params.id;
-
-  // Validate request
-
-  // Access Data
-
-  const { isSuccess, result, message } = await read(id, variant);
-  if (!isSuccess) return res.status(404).json({ message });
-
-  // Response to request
-  res.status(200).json(result);
-};
-
-const postDrugsController = async (req, res) => {
-  const record = req.body;
-  // Validate request
-
-  // Access Data
-
-  const { isSuccess, result, message } = await create(record);
-  if (!isSuccess) return res.status(404).json({ message });
-
-  // Response to request
-  res.status(201).json(result);
-};
-
-const putDrugsController = async (req, res) => {
-  const id = req.params.id;
-  const record = req.body;
-
-  // Validate request
-
-  // Access Data
-
-  const { isSuccess, result, message } = await update(record, id);
-  if (!isSuccess) return res.status(404).json({ message });
-
-  // Response to request
-  res.status(200).json(result);
-};
-
-const deleteDrugsController = async (req, res) => {
-  // Validate request
-  const id = req.params.id;
-
-  // Access Data
-  const { isSuccess, result, message } = await _delete(id);
-  if (!isSuccess) return res.status(404).json({ message });
-
-  // Response to request
-  res.status(204).json(message);
-};
+const controller = new Controller(accessor);
 
 // Endpoints -------------------------------------
 
-router.get("/", (req, res) => getDrugsController(req, res, null));
-router.get("/:id", (req, res) => getDrugsController(req, res, null));
-router.post("/", postDrugsController);
-router.put("/:id", putDrugsController);
-router.delete("/:id", deleteDrugsController);
+const router = Router();
+
+router.get("/", (req, res) => controller.get(req, res, null));
+router.get("/:id", (req, res) => controller.get(req, res, null));
+router.post("/", controller.post);
+router.put("/:id", controller.put);
+router.delete("/:id", controller.delete);
 
 export default router;
